requirements.txt
daily-python
pyaudio
pulsectl